+++ phamb/__init__.py
@@ -1,0 +1,8 @@
+
+
+"""Phamb - Phages from metagenomic binning
+Documentation: https://github.com/RasmussenLab/phamb
+"""
+
+__licence__ = 'MIT'
+__version__ = (1, 0, 1)
