--- workflows/crispr/Snakefile
@@ -1,191 +1,0 @@
-#!/usr/bin/env python
-import os   
-
-configfile: 'crispr/config.yaml'
-
-### These are gonna go into the config file at some Point
-VIRALGENOMES="04_annotation/annotation_summaries/VAMBV3.Viral_RF_predictions.bins.fna"
-MAGTAXONOMY="07_binannotation/bacteria/gtdbtk/mag_taxonomy.txt"
-VAMBDIR= config["vambdir"]
-CHECKVDIR= config["checkvdir"]
-
-THREADS=20
-
-
-### Read Binids to be parsed
-BINDIR='07_binannotation/bacteria' 
-BINIDS = []
-CHECKMFILE = config["checkm_file"]
-with open(CHECKMFILE,'r') as infile:
-    for line in infile:
-        line = line.strip().split('\t')[0]
-        if line =='bin':
-            continue
-        BINIDS.append(line)
-
-### Make List of MQNC MAGs
-with open(os.path.join(BINDIR,'MQNC_MAGS.txt'),'w') as out:
-    for binid in BINIDS:
-        out.write(os.path.join(BINDIR,binid+'.fna')+'\n')
-
-rule target: 
-    input:
-        expand("07_binannotation/bacteria/snakecrisprcasfinder/{binid}.tmp", binid=BINIDS),
-        "07_binannotation/bacteria/snakecrisprcasfinder/new.spacers.viralfragment.m6",
-        "08_crisprcas/blastn/MAGS.all.virus.m6.gz",
-        "08_crisprcas/fastani/MGXVIR.fastani.txt.gz",
-        "08_crisprcas/crispr_prophage_results/blastn_host_predictions.tsv",
-        "08_crisprcas/crispr_prophage_results/FASTANI_host_predictions.tsv"
-
-
-### This one Produces a directory of files
-rule CRISPRCAS:
-    input:
-        genome = "07_binannotation/bacteria/{binid}.fna"
-    output:
-        tmp =  "07_binannotation/bacteria/snakecrisprcasfinder/{binid}.tmp"
-    params:
-        outdir = "07_binannotation/bacteria/snakecrisprcasfinder/{binid}"
-    threads: 1
-    conda:
-         "CRISPR/envs/cctyper.yaml"
-    shell:
-        ### cctyper wants  to make the output directory, so we need to avoid Snakemake making the file by making a tmp output file
-        """
-        cctyper {input.genome} {params.outdir} --prodigal meta 
-        touch {output.tmp}
-        """
-
-
-### Map de-novo predicted spacers and also Refseq
-### Refseq spacers from https://crisprcas.i2bc.paris-saclay.fr/Home/Download 
-
-rule spacer_blastn:
-    input:
-        viralbins = VIRALGENOMES
-    output:
-        blastfile = "07_binannotation/bacteria/snakecrisprcasfinder/new.spacers.viralfragment.m6",
-        refblastfile = "07_binannotation/bacteria/snakecrisprcasfinder/RefseqNT.spacers.viralfragment.m6"
-    params:
-        refseq_spacers = "../databases/crisprcas/20210121_spacer_34.fasta"
-    threads: THREADS
-    envmodules:
-        "tools",
-        "perl/5.24.0",
-        "ncbi-blast/2.8.1+"
-    shell:
-        """
-        cat 07_binannotation/bacteria/snakecrisprcasfinder/*/spacers/*fa > 07_binannotation/bacteria/snakecrisprcasfinder/all.spacers.fna
-        makeblastdb -in {input.viralbins} -dbtype nucl
-        blastn -task blastn-short -evalue 0.001 \
-        -perc_identity 95 \
-        -num_threads {threads} \
-        -db {input.viralbins} \
-        -query 07_binannotation/bacteria/snakecrisprcasfinder/all.spacers.fna \
-        -out {output.blastfile} \
-        -outfmt '6 std qlen slen'
-
-`       blastn -task blastn-short -evalue 0.001 \
-        -perc_identity 95 \
-        -num_threads {threads} \
-        -db {input.viralbins} \
-        -query {params.refseq_spacers} \
-        -out {output.refblastfile} \
-        -outfmt '6 std qlen slen'
-        """
-
-
-### We also need to Run Both Blastn & FastANI, this does However Require that CheckV Has been RUn 
-
-rule blastn_viruses:
-    input:
-        checkv_viruses = os.path.join(CHECKVDIR,'cleaned_contigs.fna'),
-    output:
-        blastfile = "08_crisprcas/blastn/MAGS.all.virus.m6.gz"
-    params:
-        MAGgenomes_list = os.path.join(BINDIR,'MQNC_MAGS.txt'),
-        MAGgenomes = "07_binannotation/bacteria/MQNC_MAGs.fna",
-        blastparams = "-task megablast -evalue 0.001 -perc_identity 90",
-        blastfile = "08_crisprcas/blastn/MAGS.all.virus.m6"
-    threads: THREADS
-    envmodules:
-        "tools",
-        "perl/5.24.0",
-        "ncbi-blast/2.8.1+"
-    shell:
-        """
-        xargs < {params.MAGgenomes_list} cat > {params.MAGgenomes}
-        makeblastdb -in {params.MAGgenomes} -dbtype nucl
-        
-        blastn {params.blastparams} -num_threads {threads} \
-        -db {params.MAGgenomes} \
-        -query {input.checkv_viruses} \
-        -out {params.blastfile} \
-        -outfmt '6 std qlen slen'
-
-        gzip {output.blastfile}
-
-        rm {params.MAGgenomes}
-        """
-
-rule write_viruses:
-    input:
-        checkvdir = ancient( CHECKVDIR )
-    output:
-        VIRgenomes_list = "08_crisprcas/viral_genomes/viral.genomes.files.txt"
-    params:
-        VIRgenome_directory = directory( "08_crisprcas/viral_genomes" )
-    shell:
-        """
-        echo "Writing Genomes to seperate files"
-        python crispr/scripts/write_viruses.py -i {input.checkvdir} -o {params.VIRgenome_directory}
-
-        """
-
-rule fastani_viruses:
-    input:
-        checkvdir = CHECKVDIR,
-        VIRgenomes_list = "08_crisprcas/viral_genomes/viral.genomes.files.txt"
-    output:
-        fastanifile = "08_crisprcas/fastani/MGXVIR.fastani.txt.gz"
-    params:
-        fastanifile = "08_crisprcas/fastani/MGXVIR.fastani.txt",
-        VIRgenome_directory = "08_crisprcas/viral_genomes",
-        MAGgenomes_list = os.path.join(BINDIR,'MQNC_MAGS.txt')
-    threads: THREADS
-    envmodules:
-        "tools",
-        "fastani/1.1"
-    shell:
-        """
-        echo "Starting FastaANI"
-        fastANI -t {threads} \
-        --fragLen 5000 \
-        --minFrag 1 \
-        --rl {params.MAGgenomes_list} \
-        --ql {input.VIRgenomes_list} \
-        -o {params.fastanifile}
-
-        gzip {params.fastanifile}
-        """
-
-
-### Here we need to point to a lot of different Files such as MAG taxonomy, VAMB-directory, CheckV output 
-# The remaining files are already produced in the abovee steps
-
-rule MAG_Viral_connections:
-    input:
-        fastani= "08_crisprcas/fastani/MGXVIR.fastani.txt.gz",
-        blastfile = "08_crisprcas/blastn/MAGS.all.virus.m6.gz"
-    output:
-        blast_host = "08_crisprcas/crispr_prophage_results/blastn_host_predictions.tsv",
-        fastani_host = "08_crisprcas/crispr_prophage_results/FASTANI_host_predictions.tsv"
-    params:
-        VAMBDIR = VAMBDIR,
-        CHECKVDIR = CHECKVDIR,
-        MAGTAXONOMY = MAGTAXONOMY,
-    shell:
-        """
-        python crispr/scripts/crispr_prophage_summary.py -c {params.CHECKVDIR} -g {params.MAGTAXONOMY} -v {params.VAMBDIR}
-        """
-
