--- workflows/crispr/config.yaml
@@ -1,3 +1,0 @@
-checkm_file:   07_binannotation/bacteria/checkm/COPSAC.all.MQNC.MAGS
-vambdir: NA
-checkvdir: NA
